pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "iracing-oauth-client"
version = "0.1.0"
description = "A Python client for iRacing OAuth authentication using Password Limited Grant flow"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nicholas Bailey", email = "NickBaileyMA@outlook.com"}
]
keywords = ["iracing", "oauth", "api", "racing", "sim racing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/NickBaileyMA/ir-pw-limited-client"
Documentation = "https://github.com/NickBaileyMA/ir-pw-limited-client#readme"
Repository = "https://github.com/NickBaileyMA/ir-pw-limited-client"
Issues = "https://github.com/NickBaileyMA/ir-pw-limited-client/issues"

[tool.setuptools]
packages = ["iracing_oauth_client"]
